# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: GBJ SolarDAMS API — Phase 1 (Asset Registry, Maintenance, Customer Portal, Housing Developer)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Sites ───────────────────────────────────────────────────
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all PV assets
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PvAsset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a PV asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PvAssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PvAsset"
  /assets/{assetId}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get a PV asset by ID
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PV Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PvAsset"
        "404":
          description: Not found
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update a PV asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PvAssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PvAsset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete a PV asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Equipment ───────────────────────────────────────────────
...
  /developer-projects:
    get:
      operationId: listDeveloperProjects
      tags: [projects]
      summary: List housing developer projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeveloperProject"
    post:
      operationId: createDeveloperProject
      tags: [projects]
      summary: Create a housing developer project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeveloperProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperProject"
  /developer-projects/{projectId}:
    get:
      operationId: getDeveloperProject
      tags: [projects]
      summary: Get a housing developer project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperProject"
        "404":
          description: Not found
    patch:
      operationId: updateDeveloperProject
      tags: [projects]
      summary: Update a housing developer project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeveloperProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperProject"
    delete:
      operationId: deleteDeveloperProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /developer-projects/{projectId}/summary:
    get:
      operationId: getDeveloperProjectSummary
      tags: [projects]
      summary: Get installation progress summary for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"

  # ─── Housing Units ────────────────────────────────────────────
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, customerType, name, email, createdAt]
      properties:
        id:
          type: integer
        customerType:
          type: string
          enum: [residential, commercial, industrial, government]
        name:
          type: string
        identityOrRegistrationNumber:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        serviceAddress:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    CustomerInput:
      type: object
      required: [customerType, name, email]
      properties:
        customerType:
          type: string
          enum: [residential, commercial, industrial, government]
        name:
          type: string
          minLength: 1
        identityOrRegistrationNumber:
          type: string
        email:
          type: string
        phone:
          type: string
        billingAddress:
          type: string
        serviceAddress:
          type: string
    CustomerUpdate:
      type: object
      properties:
        customerType:
          type: string
          enum: [residential, commercial, industrial, government]
        name:
          type: string
        identityOrRegistrationNumber:
          type: string
        email:
          type: string
        phone:
          type: string
        billingAddress:
          type: string
        serviceAddress:
          type: string

    # ─── Site ─────────────────────────────────────────────
...
    PvAsset:
      type: object
      required: [id, siteId, assetName, assetCode, systemType, currentStatus, createdAt]
      properties:
        id:
          type: integer
        siteId:
          type: integer
        assetName:
          type: string
        assetCode:
          type: string
        systemType:
          type: string
          enum: [grid_tied, hybrid, off_grid, nem, nem2]
        installedCapacityKwp:
          type: ["number", "null"]
        acCapacityKw:
          type: ["number", "null"]
        interconnectionType:
          type: ["string", "null"]
        commissioningDate:
          type: ["string", "null"]
        designLifeYears:
          type: ["integer", "null"]
        currentStatus:
          type: string
          enum: [operational, fault, offline, under_maintenance, decommissioned]
        monitoringStatus:
          type: string
          enum: [active, inactive, not_configured]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    PvAssetInput:
      type: object
      required: [siteId, assetName, assetCode, systemType, currentStatus]
      properties:
        siteId:
          type: integer
        assetName:
          type: string
          minLength: 1
        assetCode:
          type: string
          minLength: 1
        systemType:
          type: string
          enum: [grid_tied, hybrid, off_grid, nem, nem2]
        installedCapacityKwp:
          type: number
        acCapacityKw:
          type: number
        interconnectionType:
          type: string
        commissioningDate:
          type: string
        designLifeYears:
          type: integer
        currentStatus:
          type: string
          enum: [operational, fault, offline, under_maintenance, decommissioned]
        monitoringStatus:
          type: string
          enum: [active, inactive, not_configured]
    PvAssetUpdate:
      type: object
      properties:
        siteId:
          type: integer
        assetName:
          type: string
        assetCode:
          type: string
        systemType:
          type: string
          enum: [grid_tied, hybrid, off_grid, nem, nem2]
        installedCapacityKwp:
          type: number
        acCapacityKw:
          type: number
        interconnectionType:
          type: string
        commissioningDate:
          type: string
        designLifeYears:
          type: integer
        currentStatus:
          type: string
          enum: [operational, fault, offline, under_maintenance, decommissioned]
        monitoringStatus:
          type: string
          enum: [active, inactive, not_configured]

    # ─── Equipment ────────────────────────────────────────
...
    DeveloperProject:
      type: object
      required: [id, projectName, projectCode, developerName, status, createdAt]
      properties:
        id:
          type: integer
        projectName:
          type: string
        projectCode:
          type: string
        developerName:
          type: string
        location:
          type: ["string", "null"]
        totalUnits:
          type: ["integer", "null"]
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    DeveloperProjectInput:
      type: object
      required: [projectName, projectCode, developerName, status]
      properties:
        projectName:
          type: string
          minLength: 1
        projectCode:
          type: string
          minLength: 1
        developerName:
          type: string
          minLength: 1
        location:
          type: string
        totalUnits:
          type: integer
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold]
    DeveloperProjectUpdate:
      type: object
      properties:
        projectName:
          type: string
        projectCode:
          type: string
        developerName:
          type: string
        location:
          type: string
        totalUnits:
          type: integer
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold]
    ProjectSummary:
      type: object
      required: [totalUnits, installed, commissioned, handedOver, nemSubmitted, nemApproved, outstanding]
      properties:
        totalUnits:
          type: integer
        installed:
          type: integer
        commissioned:
          type: integer
        handedOver:
          type: integer
        nemSubmitted:
          type: integer
        nemApproved:
          type: integer
        outstanding:
          type: integer

    # ─── Housing Unit ─────────────────────────────────────
...