# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * GBJ SolarDAMS API — Phase 1 (Asset Registry, Maintenance, Customer Portal, Housing Developer)
 * OpenAPI spec version: 1.0.0
 */
import type { CustomerUpdateCustomerType } from './customerUpdateCustomerType';

export interface CustomerUpdate {
  customerType?: CustomerUpdateCustomerType;
  name?: string;
  identityOrRegistrationNumber?: string;
  email?: string;
  phone?: string;
  billingAddress?: string;
  serviceAddress?: string;
}